@@ -3,10 +3,21 @@
 	"type": "module",
 	"version": "2.1.0",
 	"description": "A Svelte wrapper component for Grid.js",
+	"exports": {
+		".": {
+			"import": "./dist/index.js",
+			"require": "./dist/index.js",
+			"default": "./dist/index.js",
+			"svelte": "./dist/index.js",
+			"types": "./dist/index.d.ts"
+		}
+	},
+	"types": "./dist/index.d.ts",
+	"svelte": "./dist/index.js",
+	"main": "./dist/index.js",
+	"module": "./dist/index.js",
 	"scripts": {
-		"build": "svelte-kit package",
-		"dev:web": "svelte-kit dev",
-		"build:web": "svelte-kit build",
+		"build": "svelte-package",
 		"test": "vitest run",
 		"test:watch": "vitest watch",
 		"check": "svelte-check --tsconfig ./tsconfig.json",
@@ -21,30 +32,31 @@
 	"license": "MIT",
 	"peerDependencies": {
 		"gridjs": ">=5.0.0",
-		"svelte": ">=3.30.0"
+		"svelte": ">=5.0.0"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "next",
-		"@sveltejs/kit": "next",
-		"@sveltejs/vite-plugin-svelte": "1.0.0-next.34",
-		"@testing-library/jest-dom": "5.16.1",
-		"@testing-library/svelte": "3.0.3",
-		"@typescript-eslint/eslint-plugin": "4.31",
-		"@typescript-eslint/parser": "4.31",
-		"eslint": "7.32",
-		"eslint-config-prettier": "8.3",
-		"eslint-plugin-svelte3": "3.2",
-		"gridjs": "5.0",
-		"jsdom": "19.0.0",
-		"prettier": "2.4",
-		"prettier-plugin-svelte": "2.4",
-		"svelte": "3.44",
-		"svelte-check": "2.2",
-		"svelte-preprocess": "4.9",
-		"svelte2tsx": "0.4.13",
-		"tslib": "2.3",
-		"typescript": "4.4",
-		"vitest": "0.0.142"
+		"@sveltejs/adapter-auto": "^4.0.0",
+		"@sveltejs/kit": "^2.17.3",
+		"@sveltejs/package": "^2.3.10",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
+		"@testing-library/jest-dom": "6.6.3",
+		"@testing-library/svelte": "5.2.7",
+		"@typescript-eslint/eslint-plugin": "8.25",
+		"@typescript-eslint/parser": "8.25",
+		"eslint": "9.21",
+		"eslint-config-prettier": "10.0",
+		"eslint-plugin-svelte": "^2.45.1",
+		"gridjs": "6.2",
+		"jsdom": "26.0.0",
+		"prettier": "3.5",
+		"prettier-plugin-svelte": "3.3",
+		"svelte": "5.20",
+		"svelte-check": "4.1",
+		"svelte-preprocess": "6.0",
+		"svelte2tsx": "0.7.34",
+		"tslib": "2.8",
+		"typescript": "5.7",
+		"vitest": "3.0.7"
 	},
 	"repository": {
 		"type": "git",
