@@ -2,31 +2,39 @@
 	import type { Row, Cell } from "gridjs";
 	import Modal from "./modal.svelte";
 
-	export let row: Row;
-	export let cell: Cell;
+	interface Props {
+		row: Row;
+		cell: Cell;
+	}
 
-	$: title = `Planet ${row.cell(0).data}`;
+	let { row, cell }: Props = $props();
 
-	let showModal = false;
+	let title = $derived(`Planet ${row.cell(0).data}`);
+
+	let showModal = $state(false);
 
 	function toggleModal() {
 		showModal = !showModal;
 	}
 </script>
 
-<button class="btn" on:click={toggleModal}>See detail</button>
+<button class="btn" onclick={toggleModal}>See detail</button>
 
 <!-- for demo purpose -->
 {#if showModal}
 	<Modal on:close={toggleModal}>
-		<h2 slot="header">{title}</h2>
+		{#snippet header()}
+				<h2 >{title}</h2>
+			{/snippet}
 
 		<pre>
       {JSON.stringify(cell, null, 2)}
     </pre>
 
-		<svelte:fragment slot="action">
-			<button on:click={toggleModal}>Ok</button>
-		</svelte:fragment>
+		{#snippet action()}
+			
+				<button onclick={toggleModal}>Ok</button>
+			
+			{/snippet}
 	</Modal>
 {/if}
