@@ -1,7 +1,5 @@
-<script>
-	export let cell;
-	export let row;
-	export let column;
+<script lang="ts">
+	let { cell, row, column } = $props();
 </script>
 
 <pre data-testid="component-props">{JSON.stringify({cell, row, column}, null, 2)}</pre>
